backend/requirements.txt
# FastAPI và dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
clickhouse-driver==0.2.6
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# WebSocket
websockets==12.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP client
httpx==0.25.2

# Logging
loguru==0.7.2

# Timezone support
pytz==2023.3

# AI/LLM
google-generativeai==0.3.2

